# swagger-debug.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PhysioGest API",
    "version": "1.0.0",
    "description": "Documentação da API PhysioGest"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/financials": {
      "get": {
        "summary": "Lista todos os registros financeiros",
        "tags": [
          "Financials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de registros financeiros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Financial"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Cria um novo registro financeiro",
        "tags": [
          "Financials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "amount",
                  "date",
                  "description",
                  "userId"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ]
                  },
                  "amount": {
                    "type": "number"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "description": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registro financeiro criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Financial"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/financials/user/{userId}": {
      "get": {
        "summary": "Lista registros financeiros por usuário",
        "tags": [
          "Financials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de registros financeiros do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Financial"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/financials/{id}": {
      "get": {
        "summary": "Busca um registro financeiro por ID",
        "tags": [
          "Financials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do registro financeiro"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro financeiro encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Financial"
                }
              }
            }
          },
          "404": {
            "description": "Registro financeiro não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "put": {
        "summary": "Atualiza um registro financeiro",
        "tags": [
          "Financials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do registro financeiro"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ]
                  },
                  "amount": {
                    "type": "number"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "description": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registro financeiro atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Financial"
                }
              }
            }
          },
          "404": {
            "description": "Registro financeiro não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "summary": "Deleta um registro financeiro",
        "tags": [
          "Financials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do registro financeiro"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro financeiro deletado"
          },
          "404": {
            "description": "Registro financeiro não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Financial": {
        "type": "object",
        "required": [
          "type",
          "amount",
          "date",
          "description",
          "userId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do registro financeiro"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "description": "Tipo do registro (receita ou despesa)"
          },
          "amount": {
            "type": "number",
            "description": "Valor do registro"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "Data do registro"
          },
          "description": {
            "type": "string",
            "description": "Descrição do registro"
          },
          "userId": {
            "type": "string",
            "description": "ID do usuário que criou"
          }
        },
        "example": {
          "id": "60d5ecb74b24c72b8c8b4567",
          "type": "income",
          "amount": 150,
          "date": "2023-12-25",
          "description": "Pagamento de consulta",
          "userId": "60d5ecb74b24c72b8c8b4569"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}